R6 worker object for use as a worker with master objects generated by QueryCountMaster()

QueryCountWorker objects are worker objects at each site of a distributed QueryCount model computation

See also

Methods

Public methods

QueryCountWorker$new()

QueryCountWorker$getStateful()

QueryCountWorker$kosher()

QueryCountWorker$queryCount()

QueryCountWorker$clone()

Method new()

Create a new QueryCountWorker object.

Usage

QueryCountWorker$new(defn, data, stateful = FALSE)

Arguments

defn

the computation definition

data

the local data

stateful

the statefulness flag, default FALSE

Returns

a new QueryCountWorker object

Method getStateful()

Retrieve the value of the stateful field

Usage

QueryCountWorker$getStateful()

Method kosher()

Check if inputs and state of object are sane. For future use

Usage

QueryCountWorker$kosher()

Returns

TRUE or FALSE

Method queryCount()

Return the query count on the local data

Usage

QueryCountWorker$queryCount()

Method clone()

The objects of this class are cloneable with this method.

Usage

QueryCountWorker$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.