R6 class for SVD master object to control worker objects generated by SVDWorker()

SVDMaster objects instantiate and run a distributed SVD computation

See also

Methods

Public methods

SVDMaster$new()

SVDMaster$kosher()

SVDMaster$updateV()

SVDMaster$updateU()

SVDMaster$fixFit()

SVDMaster$reset()

SVDMaster$addSite()

SVDMaster$run()

SVDMaster$summary()

SVDMaster$clone()

Method new()

SVDMaster objects instantiate and run a distributed SVD computation

Usage

SVDMaster$new(defn, debug = FALSE)

Arguments

defn

a computation definition

debug

a flag for debugging, default FALSE

Returns

R6 SVDMaster object

Method kosher()

Check if inputs and state of object are sane. For future use

Usage

SVDMaster$kosher()

Returns

TRUE or FALSE

Method updateV()

Return an updated value for the V vector, normalized by arg

Usage

SVDMaster$updateV(arg)

Arguments

arg

the normalizing value

...

other args ignored

Returns

updated V

Method updateU()

Update U and return the updated norm of U

Usage

SVDMaster$updateU(arg)

Arguments

arg

the normalizing value

...

other args ignored

Returns

updated norm of U

Method fixFit()

Construct the residual matrix using given the V vector and d so far

Usage

SVDMaster$fixFit(v, d)

Arguments

v

the value for v

d

the value for d

Returns

result

Method reset()

Reset the computation state by initializing work matrix and set up starting values for iterating

Usage

SVDMaster$reset()

Method addSite()

Add a url or worker object for a site for participating in the distributed computation. The worker object can be used to avoid complications in debugging remote calls during prototyping.

Usage

SVDMaster$addSite(name, url = NULL, worker = NULL)

Arguments

name

of the site

url

web url of the site; exactly one of url or worker should be specified

worker

worker object for the site; exactly one of url or worker should be specified

Method run()

Run the distributed Cox model fit and return the estimates

Usage

SVDMaster$run(thr = 1e-08, max.iter = 100)

Arguments

thr

the threshold for convergence, default 1e-8

max.iter

the maximum number of iterations, default 100

Returns

a named list of V, d

Method summary()

Return the summary result

Usage

SVDMaster$summary()

Returns

a named list of V, d

Method clone()

The objects of this class are cloneable with this method.

Usage

SVDMaster$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.