Return the currently available (implemented) computations

The function availableComputations returns a list of available computations with various components. The names of this list (with no spaces) are unique canonical tags that are used throughout the package to unambiguously refer to the type of computation; web applications particularly rely on this list to instantiate objects. As more computations are implemented, this list is augmented.

availableComputations()

Value

a list with the components corresponding to a computation

desc

a textual description (25 chars at most)

definitionApp

the name of a function that will fire up a shiny webapp for defining the particular computation

workerApp

the name of a function that will fire up a shiny webapp for setting up a worker site for the particular computation

masterApp

the name of a function that will fire up a shiny webapp for setting up a master for the particular computation

makeDefinition

the name of a function that will return a data frame with appropriate fields needed to define the particular computation assuming that they are populated in a global variable. This function is used by web applications to construct a definition object based on inputs specified by the users. Since the full information is often gathered incrementally by several web applications, the inputs are set in a global variable and therefore retrieved here using the function getComputationInfo designed for the purpose

makeMaster

a function that will construct a master object for the computation given the definition and a logical flag indicating if debugging is desired

makeWorker

a function that will construct a worker object for that computation given the definition and data

See also

Examples

availableComputations()
#> $QueryCount
#> $QueryCount$desc
#> [1] "Distributed Query Count"
#> 
#> $QueryCount$definitionApp
#> [1] "defineNewQueryCountModel"
#> 
#> $QueryCount$setupWorkerApp
#> [1] "setupQueryCountWorker"
#> 
#> $QueryCount$setupMasterApp
#> [1] "setupQueryCountMaster"
#> 
#> $QueryCount$makeDefinition
#> function () 
#> {
#>     data.frame(id = getComputationInfo("id"), compType = getComputationInfo("compType"), 
#>         projectName = getComputationInfo("projectName"), projectDesc = getComputationInfo("projectDesc"), 
#>         he = getComputationInfo("he"), filterCondition = getComputationInfo("filterCondition"), 
#>         stringsAsFactors = FALSE)
#> }
#> <bytecode: 0x563be308d7e0>
#> <environment: 0x563be3093f18>
#> 
#> $QueryCount$makeMaster
#> function (defn, partyNumber, debug = FALSE) 
#> {
#>     if (!is.null(defn$he) && defn$he) {
#>         HEQueryCountMaster$new(defn = defn, partyNumber = partyNumber, 
#>             debug = debug)
#>     }
#>     else {
#>         QueryCountMaster$new(defn = defn, debug = debug)
#>     }
#> }
#> <bytecode: 0x563be308cf58>
#> <environment: 0x563be3093f18>
#> 
#> $QueryCount$makeWorker
#> function (defn, data, pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL) 
#> {
#>     if (defn$he) {
#>         HEQueryCountWorker$new(defn = defn, data = data, pubkey_bits = pubkey_bits, 
#>             pubkey_n = pubkey_n, den_bits = den_bits)
#>     }
#>     else {
#>         QueryCountWorker$new(defn = defn, data = data)
#>     }
#> }
#> <bytecode: 0x563be30901f0>
#> <environment: 0x563be3093f18>
#> 
#> 
#> $StratifiedCoxModel
#> $StratifiedCoxModel$desc
#> [1] "Stratified Cox Model"
#> 
#> $StratifiedCoxModel$definitionApp
#> [1] "defineNewCoxModel"
#> 
#> $StratifiedCoxModel$setupWorkerApp
#> [1] "setupCoxWorker"
#> 
#> $StratifiedCoxModel$setupMasterApp
#> [1] "setupCoxMaster"
#> 
#> $StratifiedCoxModel$makeDefinition
#> function () 
#> {
#>     data.frame(id = getComputationInfo("id"), compType = getComputationInfo("compType"), 
#>         projectName = getComputationInfo("projectName"), projectDesc = getComputationInfo("projectDesc"), 
#>         he = getComputationInfo("he"), formula = getComputationInfo("formula"), 
#>         stringsAsFactors = FALSE)
#> }
#> <bytecode: 0x563be308f818>
#> <environment: 0x563be3093f18>
#> 
#> $StratifiedCoxModel$makeMaster
#> function (defn, partyNumber, debug = FALSE) 
#> {
#>     if (!is.null(defn$he) && defn$he) {
#>         stop("Not implemented")
#>     }
#>     else {
#>         CoxMaster$new(defn = defn, debug = debug)
#>     }
#> }
#> <bytecode: 0x563be308ef90>
#> <environment: 0x563be3093f18>
#> 
#> $StratifiedCoxModel$makeWorker
#> function (defn, data, pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL) 
#> {
#>     if (!is.null(defn$he) && defn$he) {
#>         stop("Not implemented")
#>     }
#>     else {
#>         CoxWorker$new(defn = defn, data = data)
#>     }
#> }
#> <bytecode: 0x563be308e660>
#> <environment: 0x563be3093f18>
#> 
#> 
#> $RankKSVD
#> $RankKSVD$desc
#> [1] "Rank K SVD"
#> 
#> $RankKSVD$definitionApp
#> [1] "defineNewSVDModel"
#> 
#> $RankKSVD$setupWorkerApp
#> [1] "setupSVDWorker"
#> 
#> $RankKSVD$setupMasterApp
#> [1] "setupSVDMaster"
#> 
#> $RankKSVD$makeDefinition
#> function () 
#> {
#>     data.frame(id = getComputationInfo("id"), compType = getComputationInfo("compType"), 
#>         projectName = getComputationInfo("projectName"), projectDesc = getComputationInfo("projectDesc"), 
#>         he = getComputationInfo("he"), rank = getComputationInfo("rank"), 
#>         ncol = getComputationInfo("ncol"), stringsAsFactors = FALSE)
#> }
#> <bytecode: 0x563be3091c40>
#> <environment: 0x563be3093f18>
#> 
#> $RankKSVD$makeMaster
#> function (defn, partyNumber, debug = FALSE) 
#> {
#>     if (!is.null(defn$he) && defn$he) {
#>         stop("Not implemented")
#>     }
#>     else {
#>         SVDMaster$new(defn = defn, debug = debug)
#>     }
#> }
#> <bytecode: 0x563be30912a0>
#> <environment: 0x563be3093f18>
#> 
#> $RankKSVD$makeWorker
#> function (defn, data, pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL) 
#> {
#>     if (!is.null(defn$he) && defn$he) {
#>         stop("Not implemented")
#>     }
#>     else {
#>         SVDWorker$new(defn = defn, data = data)
#>     }
#> }
#> <bytecode: 0x563be3090970>
#> <environment: 0x563be3093f18>
#> 
#>