Given the definition identifier of an object, instantiate and store object in workspace

This function uses an identifier (defnId) to locate a stored definition in the workspace to create the appropriate object instance. The instantiated object is assigned the instanceId and saved under the dataFileName if the latter is not NULL. This instantiated object may change state between iterations when a computation executes

createNCPInstance(
  name,
  ncpId,
  instanceId,
  pubkey_bits,
  pubkey_n,
  den_bits,
  dataFileName = NULL
)

Arguments

name

identifying the NC party

ncpId

the id indicating the NCP definition

instanceId

an indentifier to use for the created instance

pubkey_bits

the public key number of bits

pubkey_n

the pubkey n

den_bits

the denominator number of bits for for rational approximations

dataFileName

a file name to use for saving the data. Typically NULL, this is only needed when one is using a single opencpu server to behave like multiple sites in which case the data file name serves to distinguish the site-specific data files. When it is NULL, the data file name is taken from the configuration settings

Value

TRUE if everything goes well