Upload a new Non-Cooperating Party (NCP) information and sites to an opencpu server

The function uploadNewNCP is really a remote version of saveNewNCP(), invoking that function on an opencpu server. This is typically done for the two NCPs participating in a computation with the list of sites. Note that sites are always a list of at least a unique name element (distinguishing the site from others) and a url element.

uploadNewNCP(defn, comp_defn, url = NULL, worker = NULL, sites)

Arguments

defn

a definition for the NCP

comp_defn

the computation definition

url

the url for the NCP. Only one of url and worker can be non-null

worker

the worker for the NCP if local. Only one of url and worker can be non-null

sites

a list of lists, each containing two items, a unique name and a (not necessarily unique) url. This is the data for the NCP!

Value

TRUE if everything goes well

See also