Create a master object to control CoxWorker worker objects

CoxMaster objects instantiate and run a distributed Cox model computation fit

See also

Methods

Public methods

CoxMaster$new()

CoxMaster$kosher()

CoxMaster$logLik()

CoxMaster$addSite()

CoxMaster$run()

CoxMaster$summary()

CoxMaster$clone()

Method new()

CoxMaster objects instantiate and run a distributed Cox model computation fit

Usage

CoxMaster$new(defn, debug = FALSE)

Arguments

defn

a computation definition

debug

a flag for debugging, default FALSE

Returns

R6 CoxMaster object

Method kosher()

Check if inputs and state of object are sane. For future use

Usage

CoxMaster$kosher()

Returns

TRUE or FALSE

Method logLik()

Return the partial log likelihood on all data for given beta parameter.

Usage

CoxMaster$logLik(beta)

Arguments

beta

the parameter vector

Returns

a named list with three components: value contains the value of the log likelihood, gradient contains the score vector, and hessian contains the estimated hessian matrix

Method addSite()

Add a url or worker object for a site for participating in the distributed computation. The worker object can be used to avoid complications in debugging remote calls during prototyping.

Usage

CoxMaster$addSite(name, url = NULL, worker = NULL)

Arguments

name

of the site

url

web url of the site; exactly one of url or worker should be specified

worker

worker object for the site; exactly one of url or worker should be specified

Method run()

Run the distributed Cox model fit and return the estimates

Usage

CoxMaster$run(control = coxph.control())

Arguments

control

parameters, same as survival::coxph.control()

Returns

a named list of beta, var, gradient, iter, and returnCode #' @description ' Return the summary of fit as a data frame

Method summary()

Usage

CoxMaster$summary()

Returns

a summary data frame columns for coef, exp(coef), ' standard error, z-score, and p-value for each parameter in the model following the same format as the survival package

Method clone()

The objects of this class are cloneable with this method.

Usage

CoxMaster$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.