Create a master object to control worker objects generated by CoxWorker

CoxMaster objects instantiate and run a distributed Cox model computation fit

CoxMaster

Format

An R6Class generator object

Methods

CoxMaster$new(defn, data, debug=FALSE)

Create a new CoxMaster object using the defn and data. The debug flag is useful for debugging

logLik(beta, ...)

Compute the partial log likelihood for all the data by aggregating the values at each site. The return value is numeric scalar with two attributes: gradient contains the score vector, and hessian contains the estimated hessian matrix

addSite(name, url = NULL, worker = NULL)

Add a worker site for participating in the distributed computation. Exactly one of url or worker should be specified

var(beta, ...)

Compute the variance of the parameter vector beta

kosher()

Check if inputs and state of object are sane. For future use

getP()

Returns the dimension of the parameter vector

run(control)

Run the fitting iterations and save the result using control object

summary()

Return a summary data frame columns for coef, exp(coef), standard error, z-score, and p-value for each parameter in the model following the same format as the survival package

See also