Create a homomorphic computation query count worker object for use with master objects generated by HEQueryCountMaster()

HEQueryCountWorker objects are worker objects at each site of a distributed query count model computation using homomorphic encryption

See also

Super class

distcomp::QueryCountWorker -> HEQueryCountWorker

Public fields

pubkey

the master's public key visible to everyone

den

the denominator for rational arithmetic

Methods

Public methods

HEQueryCountWorker$new()

HEQueryCountWorker$setParams()

HEQueryCountWorker$queryCount()

HEQueryCountWorker$clone()

Inherited methods

distcomp::QueryCountWorker$getStateful()

distcomp::QueryCountWorker$kosher()

Method new()

Create a new HEQueryMaster object.

Usage

HEQueryCountWorker$new(
  defn,
  data,
  pubkey_bits = NULL,
  pubkey_n = NULL,
  den_bits = NULL
)

Arguments

defn

the computation definition

data

the data which is usually the list of sites

pubkey_bits

the number of bits in public key

pubkey_n

the n for the public key

den_bits

the number of bits in the denominator (power of 2) used in rational approximations

Returns

a new HEQueryMaster object

Method setParams()

Set some parameters for homomorphic computations

Usage

HEQueryCountWorker$setParams(pubkey_bits, pubkey_n, den_bits)

Arguments

pubkey_bits

the number of bits in public key

pubkey_n

the n for the public key

den_bits

the number of bits in the denominator (power of 2) used in rational approximations

Method queryCount()

Run the query count on local data and return the appropriate encrypted result to the party

Usage

HEQueryCountWorker$queryCount(partyNumber, token)

Arguments

partyNumber

the NCP party number (1 or 2)

token

a token to use for identifying parts of the same computation for NCP1 and NCP2

Returns

the count as a list of encrypted items with components int and frac

Method clone()

The objects of this class are cloneable with this method.

Usage

HEQueryCountWorker$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.